feat: ProfileCard for Partners page
