+++ campus/package.json
@@ -17,8 +17,8 @@
   "devDependencies": {
     "@babel/core": "7.17.5",
     "@babel/eslint-parser": "7.17.0",
-    "@babel/plugin-syntax-flow": "^7.17.12",
-    "@babel/plugin-transform-react-jsx": "^7.17.12",
+    "@babel/plugin-syntax-flow": "^7.18.6",
+    "@babel/plugin-transform-react-jsx": "^7.18.6",
     "@babel/plugin-transform-runtime": "7.17.0",
     "@babel/preset-env": "7.16.11",
     "@babel/preset-react": "7.16.7",
@@ -48,7 +48,7 @@
     "prettier": "2.5.1",
     "prettier-eslint": "13.0.0",
     "prop-types": "^15.8.1",
-    "react-is": "^18.1.0",
+    "react-is": "^18.2.0",
     "react-native": "^0.68.2",
     "sass": "1.49.9",
     "tailwindcss": "3.0.23",
@@ -81,7 +81,7 @@
     "react-router-dom": "^6.3.0",
     "single-spa-react": "3.2.0",
     "styled-components": "5.3.3",
-    "vite": "^2.9.9",
+    "vite": "^2.9.13",
     "vite-plugin-dynamic-import": "^0.7.1"
   },
   "resolutions": {
